requirements.txt
numpy==1.16.1
pandas==0.24.0
scikit-learn==0.20.2
dependency-injector==3.14.4
Click==7.0
Sphinx==1.8.4
tinydb
pillow